refactor(#4016): one split-separator decision; string-ops now BELOW its cap
The per-function ceiling (#3400 / R-FUNC) failed on
`compileNativeStringMethodCall` (+18 on a 1,135-line function) even after the
subsystem extraction, and it was right to: only the NEW arm had moved, while the
decision it belongs to was still split across the god-function — the
undefined-separator arm (#2161 B2) 50 lines above the string-like arm, and the
ToString arm below it.

So the whole §22.1.3.23 step-2 separator decision moves together, behind one
`tryCompileStandaloneSplitSeparator` entry point that declines for a string-like
separator so the byte-identical existing arm still handles that case. Order is
preserved: undefined is still tested before the string-like arm, and the
ToString arm only ever applied when the string-like one did not.

  regexp-standalone.ts   4,261 cap:  4,516 (+255) -> 4,280 (+19)   allowance
  string-ops.ts          3,795 cap:  3,850  (+55) -> 3,755 (-40)   BELOW cap,
                                                                   allowance
                                                                   dropped
  string-search-value.ts                          ->   391         new
  compileNativeStringMethodCall  1,135 cap: 1,153 (+18) -> under

`string-ops.ts` ends up smaller than before this change, and the per-function
gate passes with no allowance at all; only the `regexp-standalone.ts` seam still
needs one.

Re-measured after the move rather than assumed — it relocated a PRE-EXISTING arm
this issue never set out to touch:

  goal-scope population   43:  0 -> 35   (search 15/15, match 11/11, split 9/9)
  all-official population 99:  0 -> 47
  regression guard       166:  166 -> 166, 0 regressed
  unit tests                   20/20 solo

Worth recording: three independent budget gates (LOC-regrowth, per-function
ceiling, oracle ratchet) each rejected a different shortcut here, and following
all three produced a better decomposition than the design I started with.

tsc clean; prettier + biome clean; LOC, func-budget, oracle-ratchet and
update-issues gates green against `upstream/main` as base (#4002).

✓
